@@ -20,10 +20,12 @@ import NIOPosix
 import OrderedCollections
 
 import struct Foundation.Date
-import class Foundation.FileHandle
-import class Foundation.Process
 import class Foundation.ProcessInfo
 import struct Foundation.URL
+
+#if canImport(Foundation.Process)
+import class Foundation.Process
+#endif
 
 // ==== ----------------------------------------------------------------------------------------------------------------
 // MARK: Code executing on each specific process/node
@@ -115,6 +117,10 @@ extension MultiNodeTestKitRunnerBoot {
     }
 
     func runMultiNodeTest(suite: String, testName: String, binary: String) async throws -> InterpretedRunResult {
+        #if !canImport(Foundation.Process)
+        struct UnsupportedPlatform: Error {}
+        throw UnsupportedPlatform()
+        #else
         let elg = MultiThreadedEventLoopGroup(numberOfThreads: 1)
         defer {
             try! elg.syncShutdownGracefully()
@@ -218,8 +224,10 @@ extension MultiNodeTestKitRunnerBoot {
 
         // all nodes passed okey
         return .passedAsExpected
+        #endif
     }
 
+    #if canImport(Foundation.Process)
     private func startTestTimeoutReaperTask(nodeName: String, process: Process, settings: MultiNodeTestSettings) -> Task<Void, Never> {
         Task.detached {
             try? await Task.sleep(until: .now + settings.execRunHardTimeout, clock: .continuous)
@@ -235,6 +243,7 @@ extension MultiNodeTestKitRunnerBoot {
             process.terminate()
         }
     }
+    #endif
 }
 
 struct NodeInterpretedRunResult {
