@@ -21,13 +21,21 @@ import OrderedCollections
 
 import struct Foundation.Date
 import class Foundation.FileHandle
-import class Foundation.Process
 import struct Foundation.URL
+
+#if canImport(Foundation.Process)
+import class Foundation.Process
+#endif
 
 @main
 struct MultiNodeTestKitRunnerBoot {
     static func main() async throws {
+        #if !canImport(Foundation.Process)
+        print("Platform does not support multi-node tests")
+        throw ExitCode.failure
+        #else
         try await Self().run()
+        #endif
     }
 
     struct TestFilter {
